+++ src/cli/utils/getEndTime.ts
