+++ src/cli/utils/getProjectDir.ts
