+++ src/cli/utils/printLog.ts
