+++ src/core/services/prepareTwigConfiguration.ts
