+++ src/core/utils/getFileType.ts
