+++ src/core/utils/loadTranslation.ts
