+++ src/definition/defines.ts
@@ -4,3 +4,5 @@ export const ENV_DEVELOPMENT: Environment = 'dev';
 export const ENV_PRODUCTION: Environment = 'prod';
 
 export const LANG_DEFAULT: string = 'cs';
+
+export const CONFIG_FILE_NAME: string = 'tateru.config.json';
